Nitpicks
Did you know: casting an enum to itself inflates stack?  What the fuck, MWCC?

+++ src/dvd/dvdfs.c
@@ -544,7 +544,7 @@ BOOL DVDSeekAsyncPrio(DVDFileInfo* fileInfo, s32 offset, DVDCallback callback, s
 	}
 
 	fileInfo->callback = callback;
-	DVDSeekAbsAsyncPrio(&fileInfo->cBlock, (u32)(char*)fileInfo->startAddr + offset, cbForSeekAsync, prio);
+	DVDSeekAbsAsyncPrio(&fileInfo->cBlock, (char*)fileInfo->startAddr + offset, cbForSeekAsync, prio);
 	return TRUE;
 }
